requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
moviepy==1.0.3
openai==1.12.0
pydub==0.25.1
python-docx==1.1.0
azure-storage-blob==12.19.0
python-dotenv==1.0.1 